Test ImageCommentsMapper in isolation
