+++ EssentialFeed/EssentialFeed/Image Comments API/ImageCommentsMapper.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-final class ImageCommentsMapper {
+final public class ImageCommentsMapper {
     
     private struct Root: Decodable {
         
@@ -28,7 +28,7 @@ final class ImageCommentsMapper {
         }
     }
     
-    static func map(_ data: Data, _ response: HTTPURLResponse) throws -> [ImageComment] {
+    public static func map(_ data: Data, _ response: HTTPURLResponse) throws -> [ImageComment] {
         let decoder = JSONDecoder()
         decoder.dateDecodingStrategy = .iso8601
