--- EssentialFeed/EssentialFeedTests/LoadImageCommentsFromRemoteUseCase.swift
@@ -1,113 +1,0 @@
-import XCTest
-import EssentialFeed
-
-final class LoadImageCommentsFromRemoteUseCase: XCTestCase {
-    
-    func test_load_deliversErrorOnNon2xxHTTPResponse() {
-        let (sut, client) = makeSUT()
-        
-        let samples = [199, 150, 300, 400, 500]
-        samples.enumerated().forEach { index, code in
-            expect(sut, toCompleteWith: failure(.invalidData), when: {
-                client.complete(withStatusCode: code, data: makeItemsJSON([]), at: index)
-            })
-        }
-    }
-    
-    func test_load_deliversErrorOnInvalidJSON() {
-        let (sut, client) = makeSUT()
-        expect(sut, toCompleteWith: failure(.invalidData), when: {
-            let invalidJSON = Data("invalid json".utf8)
-            client.complete(withStatusCode: 200, data: invalidJSON)
-        })
-    }
-
-    func test_load_deliversNoItemsOn2xxHTTPResponseWithEmptyJSON() {
-        let (sut, client) = makeSUT()
-        
-        let samples = [200, 201, 250, 280, 299]
-        samples.enumerated().forEach { index, code in
-            expect(sut, toCompleteWith: .success([])) {
-                let emptyJSON = makeItemsJSON([])
-                client.complete(withStatusCode: code, data: emptyJSON, at: index)
-            }
-        }
-    }
-    
-    func test_load_deliversItemsOn2xxHTTPResponseWithJSONItems() {
-        let (sut, client) = makeSUT()
-
-        let item1 = makeImageComment(
-            message: "a message",
-            createdAt: (Date(timeIntervalSince1970: 1645653600), "2022-02-23T22:00:00Z"),
-            username: "a username")
-        let item2 = makeImageComment(
-            message: "another message",
-            createdAt: (Date(timeIntervalSince1970: 1392890400), "2014-02-20T10:00:00Z"),
-            username: "another username")
-        
-        let samples = [200, 201, 250, 280, 299]
-        samples.enumerated().forEach { index, code in
-            expect(sut, toCompleteWith: .success([item1.model, item2.model])) {
-                let json = makeItemsJSON([item1.json, item2.json])
-                client.complete(withStatusCode: code, data: json, at: index)
-            }
-        }
-    }
-    
-    // MARK: - Helpers
-    
-    private func makeSUT(url: URL = URL(string: "http://a-url.com/")!, file: StaticString = #filePath, line: UInt = #line) -> (sut: RemoteImageCommentsLoader, client: HTTPClientSpy) {
-        let client = HTTPClientSpy()
-        let sut = RemoteImageCommentsLoader(client: client, url: url)
-        
-        trackForMemoryLeaks(sut)
-        trackForMemoryLeaks(client)
-        
-        return (sut, client)
-    }
-    
-    private func failure(_ error: RemoteImageCommentsLoader.Error) -> RemoteImageCommentsLoader.Result {
-        return .failure(error)
-    }
-    
-    private func makeImageComment(message: String, createdAt: (date: Date, iso8601String: String), username: String) -> (model: ImageComment, json: [String: Any]) {
-        let model = ImageComment(id: UUID(), message: message, createdAt: createdAt.date, username: username)
-        let json: [String: Any] = [
-            "id": model.id.uuidString,
-            "message": message,
-            "created_at": createdAt.iso8601String,
-            "author": [
-                "username": username
-            ]
-        ]
-        return (model, json)
-    }
-    
-    private func makeItemsJSON(_ items: [[String: Any]]) -> Data {
-        try! JSONSerialization.data(withJSONObject: ["items": items])
-    }
-    
-    private func expect(_ sut: RemoteImageCommentsLoader, toCompleteWith expectedResult: RemoteImageCommentsLoader.Result, when action: () -> (), file: StaticString = #filePath, line: UInt = #line) {
-        
-        let exp = expectation(description: "Wait for load completion")
-        
-        sut.load { receivedResult in
-            
-            switch (receivedResult, expectedResult) {
-            case let (.success(receivedItems), .success(expectedItems)):
-                XCTAssertEqual(receivedItems, expectedItems, file: file, line: line)
-            case let (.failure(receivedError as RemoteImageCommentsLoader.Error), .failure(expectedError as RemoteImageCommentsLoader.Error)):
-                XCTAssertEqual(receivedError, expectedError, file: file, line: line)
-            default:
-                XCTFail("Expected result \(expectedResult), got \(receivedResult) instead", file: file, line: line)
-            }
-            
-            exp.fulfill()
-        }
-        
-        action()
-        
-        wait(for: [exp], timeout: 1.0)
-    }
-}
